requirements.txt
gradio==5.12.0
transformers>=4.50.0
torch==2.5.1
bitsandbytes==0.45.0
accelerate==1.2.0
pillow==10.4.0
fpdf2==2.8.0
pandas==2.2.3
matplotlib==3.9.4
python-dotenv==1.0.1
huggingface_hub>=0.25.0
openai-whisper>=20231117
